+++ design/src/androidTest/kotlin/uk/govuk/app/design/ui/component/ButtonsTest.kt
@@ -1,4 +1,4 @@
-package uk.govuk.app.design.ui.components
+package uk.govuk.app.design.ui.component
 
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.test.assertIsDisplayed
